refactor: remove `screen` module
This name does not explain which codes should be included.
