--- kernel/src/graphics/screen/mod.rs
@@ -1,5 +1,0 @@
-// SPDX-License-Identifier: GPL-3.0-or-later
-
-pub mod layer;
-pub mod log;
-pub mod writer;
