+++ kernel/src/graphics/writer.rs
